Переписал на модули с помощью CommonJS
