+++ src/index.html
@@ -126,10 +126,6 @@
     </div>
   </article>
 </template>
-
-<script src="js/form.js"></script>
-<script src="js/game.js"></script>
-<script src="js/reviews.js"></script>
 <script src="js/main.js"></script>
 </body>
 </html>
