+++ src/js/main.js
@@ -1,9 +1,14 @@
 'use strict';
 
 (function() {
-  var game = new window.Game(document.querySelector('.demo'));
+  var Game = require('./game');
+  var form = require('./form');
+  var reviewList = require('./review');
+
+  var game = new Game(document.querySelector('.demo'));
+
   game.initializeLevelAndStart();
-  game.setGameStatus(window.Game.Verdict.INTRO);
+  game.setGameStatus(Game.Verdict.INTRO);
 
   var formOpenButton = document.querySelector('.reviews-controls-new');
 
@@ -11,13 +16,21 @@
   formOpenButton.onclick = function(evt) {
     evt.preventDefault();
 
-    window.form.open(function() {
-      game.setGameStatus(window.Game.Verdict.PAUSE);
+    form.open(function() {
+      game.setGameStatus(Game.Verdict.PAUSE);
       game.setDeactivated(true);
     });
   };
 
-  window.form.onClose = function() {
+  form.onClose = function() {
     game.setDeactivated(false);
+  };
+
+  /**
+   * Запуск рендера отзывов
+   * @param {Array} data
+   */
+  window.renderReviews = function(data) {
+    reviewList.render(data);
   };
 })();
